refactor:Response Body Formatter 적용
